Add better debug console output

+++ Book_Invoicings_Example/SOLID/src/Book.java
@@ -5,10 +5,10 @@ class Book {
 	String name;
 	String authorName;
 	int year;
-	int price;
+	float price;
 	String isbn;
  
-	public Book(String name, String authorName, int year, int price, String isbn) {
+	public Book(String name, String authorName, int year, float price, String isbn) {
 		this.name = name;
 		this.authorName = authorName;
 		this.year = year;
